requirements.txt
authy==2.2.3
awsebcli==3.14.11
blessed==1.15.0
botocore==1.12.91
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
Django==2.1.5
djangorestframework==3.9.1
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.7
jmespath==0.9.3
jsonschema==2.6.0
pathspec==0.5.5
phonenumbers==8.10.5
psycopg2-binary==2.7.7
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.54.0
